+++ app/controllers/players.js
@@ -165,6 +165,8 @@ var sendConfig = function (player, group, periodic) {
     retObj.emergencyMessage = group.emergencyMessage || {enable: false};
     retObj.combineDefaultPlaylist = group.combineDefaultPlaylist || false;
     retObj.playAllEligiblePlaylists = group.playAllEligiblePlaylists || false;
+    retObj.shuffleContent = group.shuffleContent || false;
+    retObj.alternateContent = group.alternateContent || false;
     retObj.urlReloadDisable =  group.urlReloadDisable || false;
     retObj.loadPlaylistOnCompletion =  group.loadPlaylistOnCompletion || false;
     //if (!pipkgjson)
@@ -190,7 +192,9 @@ var sendConfig = function (player, group, periodic) {
     retObj.currentTime = Date.now();
     var socketio = (player.newSocketIo?newSocketio:oldSocketio);
     socketio.emitMessage(player.socket, 'config', retObj);
+    return retObj;
 }
+exports.sendConfig=sendConfig;
 
 //Load a object
 exports.loadObject = function (req, res, next, id) {