+++ public/app/js/app-pc.js
@@ -9,6 +9,7 @@ angular.module('piServerApp', [
     'piConfig',
     'piIndex.controllers',
     'piGroups.controllers',
+    'dashboard.controllers',
     'piAssets.controllers',
     'piAssets.services',
     'piPlayers.services',
@@ -57,6 +58,26 @@ angular.module('piServerApp', [
                     },
                     "list": {
                         templateUrl: '/app/partials/players.html'
+                    }
+                }
+            })
+
+            .state("home.dashboard",{
+                url: "dashboard",
+                views: {
+                    "main": {
+                        templateUrl: '/app/partials/dashboard.html',
+                        controller: 'DashboardCtrl'
+                    }
+                }
+            })
+
+            .state("home.dashboard_players", {
+                url: "dashboard/players?groupName&locationName&currentPlaylist&version&bucket",
+                views: {
+                    "main": {
+                        templateUrl: '/app/partials/players.html',
+                        controller:'ServerPlayerCtrl'
                     }
                 }
             })