+++ public/app/js/directives/pi.js
@@ -105,7 +105,7 @@ directive('nodeimsFileUpload', ['fileUploader','piUrls', function(fileUploader, 
                 tElement.attr("multiple", "multiple");
             }        
             if (!tAttrs.maxFileSizeMb) {
-                tAttrs.maxFileSizeMb = 3000;
+                tAttrs.maxFileSizeMb = 5000;
             }        
             return function postLink(scope, el, attrs, ctl) {                
                 scope.files = [];
@@ -119,7 +119,7 @@ directive('nodeimsFileUpload', ['fileUploader','piUrls', function(fileUploader, 
                     
                     scope.files = [];
                     scope.maxFiles = scope.maxFiles || 10;
-                    scope.maxFileSizeMb = scope.maxFileSizeMb || 3000;
+                    scope.maxFileSizeMb = scope.maxFileSizeMb || 5000;
                     var tooBig = [];
                     if (e.target.files.length > scope.maxFiles) {
                         raiseError(e.target.files, 'TOO_MANY_FILES', "Cannot upload " + e.target.files.length + " files, maxium allowed is " + scope.maxFiles);