+++ public/app/js/services/pi.js
@@ -4,6 +4,9 @@
 
 angular.module('pisignage.services',[]).
 
+factory('selectedLabel', function () {
+    return ({selectedLabel: null, labelsCount: {}})
+}).
     //https://github.com/logicbomb/lvlFileUpload
     factory('fileUploader', ['$rootScope', '$q', function($rootScope, $q) {
         var xhr, completeTransferDone;